+++ corehq/apps/case_search/templates/case_search/csql_fixture_configuration.html
