+++ corehq/apps/domain/deletion.py
@@ -468,8 +468,8 @@ DOMAIN_DELETE_OPERATIONS = [
     ModelDeletion('couchforms', 'UnfinishedSubmissionStub', 'domain'),
     ModelDeletion('couchforms', 'UnfinishedArchiveStub', 'domain'),
     ModelDeletion('fixtures', 'LookupTable', 'domain'),
-    ModelDeletion('fixtures', 'CSQLFixtureExpression', 'domain'),
-    ModelDeletion('fixtures', 'CSQLFixtureExpressionLog', 'expression__domain'),
+    ModelDeletion('case_search', 'CSQLFixtureExpression', 'domain'),
+    ModelDeletion('case_search', 'CSQLFixtureExpressionLog', 'expression__domain'),
     CustomDeletion('ucr', delete_all_ucr_tables_for_domain, []),
     ModelDeletion('domain', 'OperatorCallLimitSettings', 'domain'),
     ModelDeletion('domain', 'SMSAccountConfirmationSettings', 'domain'),
