+++ corehq/apps/fixtures/models.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from django.db.models import ForeignKey
 from functools import reduce
 from itertools import chain
 from uuid import uuid4
@@ -7,10 +6,6 @@ from uuid import uuid4
 from attrs import define, field
 from django.db import models
 from django.db.models.expressions import RawSQL
-from django.db.models.signals import post_save
-from django.db.transaction import atomic
-from django.dispatch import receiver
-from django.utils.translation import gettext as _
 
 from corehq.apps.groups.models import Group
 from corehq.sql_db.fields import CharIdField
@@ -310,50 +305,3 @@ class UserLookupTableStatus(models.Model):
         app_label = 'fixtures'
         db_table = 'fixtures_userfixturestatus'
         unique_together = ("user_id", "fixture_type")
-
-
-class CSQLFixtureExpression(models.Model):
-    domain = models.CharField(max_length=64, default='')
-    name = models.CharField(max_length=64, null=False)
-    csql = models.CharField(null=False)
-    date_created = models.DateTimeField(auto_now_add=True)
-    last_modified = models.DateTimeField(auto_now=True)
-    deleted = models.BooleanField(default=False)
-
-    @classmethod
-    def by_domain(cls, domain):
-        return cls.objects.filter(domain=domain, deleted=False)
-
-    @atomic
-    def soft_delete(self):
-        self.deleted = True
-        self.save()
-        CSQLFixtureExpressionLog.objects.create(
-            expression=self,
-            action=CSQLFixtureExpressionLog.Action.DELETE.value,
-        )
-
-
-class CSQLFixtureExpressionLog(models.Model):
-    class Action(models.TextChoices):
-        CREATE = 'CR', _('Create')
-        DELETE = 'DE', _('Delete')
-        UPDATE = 'UP', _('Update')
-
-    expression = ForeignKey(CSQLFixtureExpression, on_delete=models.CASCADE)
-    date = models.DateTimeField(auto_now_add=True)
-    action = models.CharField(max_length=2, choices=Action.choices, null=False)
-    name = models.CharField(max_length=64, default='')
-    csql = models.TextField(default='')
-
-
-@receiver(post_save, sender=CSQLFixtureExpression)
-def after_save(sender, instance, created, **kwargs):
-    updated_or_created = (CSQLFixtureExpressionLog.Action.CREATE.value if created
-                          else CSQLFixtureExpressionLog.Action.UPDATE.value)
-    CSQLFixtureExpressionLog.objects.create(
-        expression=instance,
-        action=updated_or_created,
-        name=instance.name,
-        csql=instance.csql
-    )
