+++ corehq/apps/fixtures/views.py
@@ -12,7 +12,6 @@ from django.http import (
     Http404,
     HttpResponseBadRequest,
     HttpResponseRedirect,
-    HttpResponse,
     JsonResponse,
 )
 from django.http.response import HttpResponseServerError
@@ -25,7 +24,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 from django.views.generic.base import TemplateView
 
-from corehq.apps.hqwebapp.decorators import use_bootstrap5, waf_allow
+from corehq.apps.hqwebapp.decorators import waf_allow
 from dimagi.utils.decorators.view import get_file
 from dimagi.utils.logging import notify_exception
 from dimagi.utils.web import get_url_base, json_response
@@ -44,7 +43,7 @@ from corehq.apps.fixtures.exceptions import (
     FixtureUploadError,
 )
 from corehq.apps.fixtures.fixturegenerators import item_lists_by_domain
-from corehq.apps.fixtures.models import (LookupTable, LookupTableRow, CSQLFixtureExpression,
+from corehq.apps.fixtures.models import (LookupTable, LookupTableRow,
     TypeField)
 from corehq.apps.fixtures.tasks import (
     async_fixture_download,
@@ -573,77 +572,3 @@ def fixture_metadata(request, domain):
     Returns list of fixtures and metadata needed for itemsets in vellum
     """
     return json_response(item_lists_by_domain(domain))
-
-
-@method_decorator([
-    use_bootstrap5,
-    toggles.MODULE_BADGES.required_decorator(),
-    require_can_edit_fixtures,
-], name='dispatch')
-class CSQLFixtureExpressionView(BaseDomainView):
-    urlname = 'csql_fixture_configuration'
-    page_title = _('CSQL Fixture Confguration')
-    template_name = 'fixtures/csql_fixture_configuration.html'
-
-    @property
-    def section_url(self):
-        return reverse(self.urlname, args=[self.domain])
-
-    def all_module_badge_configurations(self):
-        return CSQLFixtureExpression.by_domain(self.domain)
-
-    @property
-    def page_context(self):
-        return {
-            'save_url': reverse(self.urlname, args=[self.domain]),
-            'csql_fixture_configurations':
-                list(self.all_module_badge_configurations().values('id', 'name', 'csql')),
-        }
-
-    def _post_response(self, message, div_class):
-        return HttpResponse((f'<div class="alert {div_class}">' + _(message) + '</div>'),
-                            content_type='text/html')
-
-    @atomic
-    def post(self, request, *args, **kwargs):
-        try:
-            data = request.POST
-            ''' Post data format is:
-                {
-                    'name': ['name1', 'name2', 'name3'],
-                    'id': ['1', '2', ''],  # empty string ID means new expression, missing means delete
-                    'csql': ['asdf', 'asdfg', 'asdfgh'],}
-                }
-            '''
-
-            ids_list = data.getlist('id')
-            name_list = data.getlist('name')
-            csql_list = data.getlist('csql')
-
-            name_list_without_blanks = [name for name in name_list if name]
-            if len(name_list_without_blanks) != len(set(name_list_without_blanks)):
-                return self._post_response(
-                    "Configuration not updated, two expressions cannot have the same name.", 'alert-warning')
-            for i in range(0, len(name_list)):
-                if not name_list[i] or not csql_list[i]:
-                    return self._post_response("Configuration not updated, some fields are blank.",
-                                               'alert-warning')
-
-            id_list_without_blanks = [_id for _id in ids_list if _id]
-            for expression in CSQLFixtureExpression.by_domain(self.domain).exclude(id__in=id_list_without_blanks):
-                expression.soft_delete()
-
-            for _id, name, csql in zip(ids_list, name_list, csql_list):
-                if _id:
-                    module_badge_configuration = CSQLFixtureExpression.by_domain(self.domain).get(id=_id)
-                    if name != module_badge_configuration.name or csql != module_badge_configuration.csql:
-                        module_badge_configuration.name = name
-                        module_badge_configuration.csql = csql
-                        module_badge_configuration.save()
-                else:
-                    CSQLFixtureExpression.objects.create(
-                        domain=self.domain, name=name, csql=csql)
-            return self._post_response("Fixture confugration updated!", 'alert-success')
-        except Exception as e:
-            notify_exception(request, message=str(e))
-            return self._post_response("Configuration not updated, unknown error occurred.", 'alert-danger')
